SmartAI/AI/requirements.txt
torch
torch-geometric
torch-cluster 
torch-scatter 
torch-sparse 
torch-spline-conv 
torch-geometric-temporal
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
networkx
hydra-core
omegaconf
tensorboard
umap-learn
joblib
pytest